# speedup import time by caching numba vectorizes