@FedericoTartarini 2025-01-07
Thank you very much for your help and for contributing to pythermalcomfort